# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/v1/business-subcategories/sector/{sectorId}": {
            "get": {
                "description": "Get business subcategories filtered by sector ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Business"
                ],
                "summary": "Get business subcategories by sector",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Sector ID",
                        "name": "sectorId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.BaseResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/models.BusinessSubcategory"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
        "/api/v1/businesses/by-sector": {
            "get": {
                "description": "Get all businesses in a specific sector (e.g., technology, financial)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Business"
                ],
                "summary": "Get businesses by sector",
                "parameters": [
                    {
                        "enum": [
                            "retail",
                            "wholesale",
                            "fashion",
                            "beauty",
                            "food",
                            "health",
                            "agriculture",
                            "construction",
                            "real_estate",
                            "transport",
                            "logistics",
                            "hospitality",
                            "tourism",
                            "education",
                            "professional",
                            "financial",
                            "technology",
                            "telecom",
                            "energy",
                            "manufacturing",
                            "mining",
                            "automotive",
                            "entertainment",
                            "media",
                            "sports",
                            "creative",
                            "community",
                            "environment",
                            "security",
                            "cleaning",
                            "veterinary"
                        ],
                        "type": "string",
                        "description": "Sector",
                        "name": "sector",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "maximum": 100,
                        "minimum": 1,
                        "type": "integer",
                        "default": 20,
                        "description": "Page size",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.BaseResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "additionalProperties": true
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.BaseResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.BaseResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/businesses/by-type": {
            "get": {
                "description": "Get all businesses of a specific type (e.g., private_company, sole_proprietorship)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Business"
                ],
                "summary": "Get businesses by business type",
                "parameters": [
                    {
                        "enum": [
                            "sole_proprietorship",
                            "sole_trader",
                            "partnership",
                            "limited_partnership",
                            "private_company",
                            "public_company",
                            "limited_by_guarantee",
                            "cooperative",
                            "sacco",
                            "ngo",
                            "cbo",
                            "trust",
                            "foundation",
                            "faith_based",
                            "franchise",
                            "epz",
                            "special_economic",
                            "state_corporation",
                            "government_agency"
                        ],
                        "type": "string",
                        "description": "Business type",
                        "name": "business_type",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "maximum": 100,
                        "minimum": 1,
                        "type": "integer",
                        "default": 20,
                        "description": "Page size",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.BaseResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "additionalProperties": true
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.BaseResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.BaseResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.BusinessSector": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "display_name": {
                    "description": "Display: \"Financial Services\"",
                    "type": "string"
                },
                "icon": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_active": {
                    "type": "boolean"
                },
                "name": {
                    "description": "Constant: \"financial\"",
                    "type": "string"
                },
                "sort_order": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "models.BusinessSubcategory": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "display_name": {
                    "description": "✅ ADD THIS",
                    "type": "string"
                },
                "icon": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_active": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "sector": {
                    "description": "Relationship",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.BusinessSector"
                        }
                    ]
                },
                "sector_id": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "response.BaseResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "error": {
                    "$ref": "#/definitions/response.ErrorPayload"
                },
                "message": {
                    "type": "string"
                },
                "pagination": {
                    "$ref": "#/definitions/response.PaginationMeta"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "response.ErrorPayload": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "details": {}
            }
        },
        "response.PaginationMeta": {
            "type": "object",
            "properties": {
                "hasMore": {
                    "type": "boolean"
                },
                "limit": {
                    "type": "integer"
                },
                "offset": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
    }
}